notification-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic[email]
kafka-python
aiosmtplib
pydantic-settings
aiokafka